Compute GEV Spectrum Diagnostics

Computes various diagnostic statistics for a vector of Generalized Eigenvalues (GEV).

Usage

compute_gev_spectrum_diagnostics(Lambda_GEV, lambda_max_thresh)

Arguments

Lambda_GEV

A numeric vector of eigenvalues obtained from GEV.

lambda_max_thresh

A numeric threshold used to categorize eigenvalues. Eigenvalues with absolute value less than this are considered 'retained' or 'stable'.

Value

A list containing the following diagnostic statistics: \itemize \itemn_eigenvalues: Total number of eigenvalues. \itemmin_eigenvalue: Minimum eigenvalue. \itemmax_eigenvalue: Maximum eigenvalue. \itemmean_eigenvalue: Mean of eigenvalues. \itemmedian_eigenvalue: Median of eigenvalues. \itemsd_eigenvalue: Standard deviation of eigenvalues. \itemn_below_thresh: Number of eigenvalues with absolute value < lambda_max_thresh. \itemprop_below_thresh: Proportion of eigenvalues with absolute value < lambda_max_thresh. \itemn_above_thresh: Number of eigenvalues with absolute value >= lambda_max_thresh. \itemprop_above_thresh: Proportion of eigenvalues with absolute value >= lambda_max_thresh.

Examples

  Lambda_GEV_sample <- c(0.1, 0.5, 0.85, 0.95, 1.2)
  compute_gev_spectrum_diagnostics(Lambda_GEV_sample, lambda_max_thresh = 0.8)
#> $n_eigenvalues
#> [1] 5
#> 
#> $min_eigenvalue
#> [1] 0.1
#> 
#> $max_eigenvalue
#> [1] 1.2
#> 
#> $mean_eigenvalue
#> [1] 0.72
#> 
#> $median_eigenvalue
#> [1] 0.85
#> 
#> $sd_eigenvalue
#> [1] 0.4280771
#> 
#> $n_below_thresh
#> [1] 2
#> 
#> $prop_below_thresh
#> [1] 0.4
#> 
#> $n_above_thresh
#> [1] 3
#> 
#> $prop_above_thresh
#> [1] 0.6
#>